1/integer值比较问题修改 2/登出删除

+++ source/kom-manager-service/src/main/java/com/transsion/store/manager/OrganizationManager.java
@@ -98,7 +98,7 @@ public class OrganizationManager {
 		for(OrganizationTreeDto org:orgList){
 			Integer orgId = org.getOrgId();  
 	    	 Integer pid = org.getParentId();
-	           if (parentId == pid) {  
+	           if (parentId.intValue() == pid.intValue()) {  
 	        	   List<OrganizationTreeDto> children = getChridenOrg(orgList, orgId); 
 	        	   org.setChildren(children);
 	        	   orgTreeDto.add(org);
